Workaround for document type handling
The quality of the workaround is very poor. It's a hack to submit the form in a
custom way whenever the document type is changed. This allows to use the form
and to explore the intended behavior.

A proper implementation will have to handle an update of the metadata subsection
when the value for the document type changes. A refresh of the form might be a
viable approach if the form can render itself for a "target_document_type" which
is somehow provided based on the selection.